Refactor Design

+++ app/src/main/java/com/stho/mayai/ui/alarms/SwipeAlarmToDelete.java
@@ -1,43 +1,77 @@
 package com.stho.mayai.ui.alarms;
 
+import android.graphics.Canvas;
+import android.graphics.Color;
+import android.graphics.drawable.ColorDrawable;
+import android.graphics.drawable.Drawable;
 import android.text.Layout;
+import android.view.View;
 
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.core.content.ContextCompat;
+import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.ItemTouchHelper;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.stho.mayai.R;
+
+import org.jetbrains.annotations.NotNull;
 
 import static androidx.recyclerview.widget.ItemTouchHelper.LEFT;
 import static androidx.recyclerview.widget.ItemTouchHelper.RIGHT;
 
 /*
-    See: https://medium.com/@zackcosborn/step-by-step-recyclerview-swipe-to-delete-and-undo-7bbae1fce27e
+    See:
+        https://medium.com/@zackcosborn/step-by-step-recyclerview-swipe-to-delete-and-undo-7bbae1fce27e
+        https://www.androidhive.info/2017/09/android-recyclerview-swipe-delete-undo-using-itemtouchhelper/
  */
 public class SwipeAlarmToDelete extends ItemTouchHelper.SimpleCallback {
 
-    private AlarmsRecyclerViewAdapter adapter;
+    private final AlarmsRecyclerViewAdapter adapter;
 
     public SwipeAlarmToDelete(AlarmsRecyclerViewAdapter adapter) {
-        super(0,ItemTouchHelper.LEFT | RIGHT);
+        super(0, ItemTouchHelper.LEFT);
         this.adapter = adapter;
     }
 
     @Override
     public boolean onMove(@NonNull RecyclerView recyclerView, @NonNull RecyclerView.ViewHolder viewHolder, @NonNull RecyclerView.ViewHolder target) {
-        return false;
+        return true;
     }
 
     @Override
     public void onSwiped(@NonNull RecyclerView.ViewHolder viewHolder, int direction) {
         int position = viewHolder.getAdapterPosition();
-        switch (direction) {
-            case RIGHT:
-                adapter.editItem(position);
-                break;
-
-            case LEFT:
-                adapter.deleteItem(position);
-                break;
+        if (direction == LEFT) {
+            adapter.deleteItem(position);
         }
+    }
+
+    @Override
+    public void onSelectedChanged(@Nullable RecyclerView.ViewHolder viewHolder, int actionState) {
+        if (viewHolder != null) {
+            final View foregroundView = ((AlarmsRecyclerViewAdapter.ViewHolder) viewHolder).viewForeground;
+            getDefaultUIUtil().onSelected(foregroundView);
+        }
+    }
+
+    @Override
+    public void onChildDrawOver(@NonNull Canvas c, @NonNull RecyclerView recyclerView, RecyclerView.ViewHolder viewHolder, float dX, float dY, int actionState, boolean isCurrentlyActive) {
+        final View foregroundView = ((AlarmsRecyclerViewAdapter.ViewHolder) viewHolder).viewForeground;
+        getDefaultUIUtil().onDrawOver(c, recyclerView, foregroundView, dX, dY, actionState, isCurrentlyActive);
+    }
+
+    @Override
+    public void clearView(@NonNull RecyclerView recyclerView, @NonNull RecyclerView.ViewHolder viewHolder) {
+        final View foregroundView = ((AlarmsRecyclerViewAdapter.ViewHolder) viewHolder).viewForeground;
+        getDefaultUIUtil().clearView(foregroundView);
+    }
+
+    @Override
+    public void onChildDraw(@NotNull Canvas c, @NotNull RecyclerView recyclerView, @NotNull RecyclerView.ViewHolder viewHolder, float dX, float dY, int actionState, boolean isCurrentlyActive) {
+        final View foregroundView = ((AlarmsRecyclerViewAdapter.ViewHolder) viewHolder).viewForeground;
+        getDefaultUIUtil().onDraw(c, recyclerView, foregroundView, dX, dY, actionState, isCurrentlyActive);
     }
 }
